% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game-attribs.R
\name{del_player}
\alias{del_player}
\title{Delete a player from the game}
\usage{
del_player(game, player)
}
\arguments{
\item{game}{An \code{alicetwist} game object}

\item{player}{The ID of the player that is
to be removed from the game}
}
\description{
Delete a player from the game
}
\keyword{internal}
