% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game-attribs.R
\name{game_fingers}
\alias{game_fingers}
\title{Get the number of fingers}
\usage{
game_fingers(game)
}
\arguments{
\item{game}{An \code{alicetwist} game object}
}
\description{
Get the number of fingers
}
\keyword{internal}
