% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game-attribs.R
\name{game_hands}
\alias{game_hands}
\title{Get the number of hands}
\usage{
game_hands(game)
}
\arguments{
\item{game}{An \code{alicetwist} game object}
}
\description{
Get the number of hands
}
\keyword{internal}
