% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game-attribs.R
\name{game_players}
\alias{game_players}
\title{Get the number of players}
\usage{
game_players(game)
}
\arguments{
\item{game}{An \code{alicetwist} game object}
}
\description{
Get the number of players
}
\keyword{internal}
