% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game-attribs.R
\name{game_split}
\alias{game_split}
\title{Is the game allowing the split move?}
\usage{
game_split(game)
}
\arguments{
\item{game}{An \code{alicetwist} game object}
}
\description{
Split move allows players who have zero
fingers left in one hand to move fingers
from other hands.
}
\keyword{internal}
