% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game-attribs.R
\name{game_status<-}
\alias{game_status<-}
\title{Set the game status}
\usage{
game_status(game, player) <- value
}
\arguments{
\item{game}{An \code{alicetwist} game object}

\item{player}{The player whose hands are to
be modified}

\item{value}{A vector containing the new
number of fingers to assign to each hand}
}
\description{
Set the game status
}
\keyword{internal}
