% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game-attribs.R
\name{game_status}
\alias{game_status}
\title{Get the game status (number of fingers alive)}
\usage{
game_status(game)
}
\arguments{
\item{game}{An \code{alicetwist} game object}
}
\value{
A matrix representing the current state
of the game (number of fingers still alive) with
the players on the rows and the hands on the columns
}
\description{
Get the game status (number of fingers alive)
}
\examples{
game <- new_game()
game
game_status(game)
}
