% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play}
\alias{play}
\title{Play a single move in the game}
\usage{
play(game, active_player, active_hand, target_player, target_hand, vector)
}
\arguments{
\item{game}{An \code{alicetwist} game object.}

\item{active_player}{The ID of the striker player.}

\item{active_hand}{The ID of the hand with which
the \code{active_player} hits the opponent's hand.}

\item{target_player}{The ID of the player who
undergoes the hit.}

\item{target_hand}{The ID of the hand which is
been hit.}

\item{vector}{Instead of using \code{active_player},
\code{active_hand}, \code{target_player} and
\code{target_hand}, a vector can be used to
specify in one shot the four IDs.}
}
\value{
A new \code{alicetwist} game object after
the given move is accomplished.
}
\description{
Pushing one hand of a player against the hand
of another player implies that the number of
fingers in the target hands becomes the sum
of the original fingers' number and the hitting
fingers' number, modulus the number of fingers
in the hands.
}
\examples{
# Create a basic game and then make a move

game <- new_game()
game

# Make a move
game_after_move <- play(game, 1, 1, 2, 1)
game_after_move

# Make a move for both the players
game_after_move <- game \%>\% play(1, 1, 2, 1) \%>\% play(2, 1, 1, 1)
game_after_move
}
\seealso{
\code{\link[alicetwist]{new_game}}
}
